requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
boto3==1.37.2
botocore==1.37.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
execnet==2.1.1
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
openai==1.65.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
responses==0.25.6
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
